requirements.txt
streamlit>=1.44.0
pandas>=2.2.0
plotly>=5.24.0
openai>=1.68.0
python-dotenv>=1.0.1
